Start replacing openraft with raft
