+++ Cargo.toml
@@ -26,8 +26,10 @@ rand = "0.8.5"
 serde_regex = "1.1.0"
 jsonl = { version="4.0.1", features=["tokio"]}
 anyhow = "1.0"
-openraft = { git = "https://github.com/datafuselabs/openraft", rev = "2262c79f5195307402e7a0994771b4152c0d10b2" }
 sled = "0.34.7"
+raft = { version = "0.6.0" }
+# Maybe remove this if we bump raft?  https://github.com/tikv/raft-rs/issues/463
+protobuf = { version = "2", features = ["bytes"] }
 
 async-trait = "0.1.36"
 clap = { version = "3.0.13", features = ["derive", "env"] }
