+++ src/garbage.rs
@@ -19,7 +19,8 @@ use crate::{
 const MINIMUM_GARBAGE_AGE: i64 = 60 * 60 * 12;
 
 async fn do_garbage_collect_phase1(app: &Arc<RegistryApp>) {
-    if app.raft.is_leader().await.is_err() {
+    //if app.raft.is_leader().await.is_err() {
+    if true {
         info!("Garbage collection: Phase 1: Not leader");
         return;
     }
