+++ src/network/mod.rs
@@ -1,4 +1,1 @@
-pub mod api;
-pub mod management;
-pub mod raft;
-pub mod raft_network_impl;
+pub mod server;
